+++ internal/app/singularity/actions_darwin.go
@@ -6,26 +6,8 @@
 package cli
 
 import (
-	"fmt"
-
 	"github.com/spf13/cobra"
 )
-
-func handleOCI(cmd *cobra.Command, u string) (string, error) {
-	return "", fmt.Errorf("unsupported on this platform")
-}
-
-func handleLibrary(u string) (string, error) {
-	return "", fmt.Errorf("unsupported on this platform")
-}
-
-func handleShub(u string) (string, error) {
-	return "", fmt.Errorf("unsupported on this platform")
-}
-
-func handleNet(u string) (string, error) {
-	return "", fmt.Errorf("unsupported on this platform")
-}
 
 // TODO: Let's stick this in another file so that that CLI is just CLI
 func execStarter(cobraCmd *cobra.Command, image string, args []string, name string) {
